% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwdata-package.R
\docType{package}
\name{cwdata-package}
\alias{cwdata}
\alias{cwdata-package}
\title{cwdata: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dea Avega Editya \email{vegatroz@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
